# :racehorse: speedup ignore handling