This change introduces a speedup for `snow status` and `snow add`. Both rely on the ignore handling, which was slow due to the call of `ignore.ignored(file, this.patterns)` for each single file. The module [micromatch](https://github.com/micromatch/micromatch) which is used under the hood, evaluated the regex for each file. Passing over a list of all files in a single shot results in the speed improvements as demonstrated below - same repo as from the initial bug report (> 40000 files):

```
$ ls -lsa | wc -l
40003

$ time /c/Users/sebastian/Documents/git/SnowFS/dist/out-tsc/snow.exe status > /c/Users/sebastian/Desktop/output.txt

real    0m2.142s
user    0m0.015s
sys     0m0.000s

$ time /c/Users/sebastian/Documents/git/SnowFS/dist/out-tsc/snow.exe add .

real    0m31.213s
user    0m0.015s
sys     0m0.000s

$ time /c/Users/sebastian/Documents/git/SnowFS/dist/out-tsc/snow.exe status > /c/Users/sebastian/Desktop/output.txt

real    0m2.502s
user    0m0.015s
sys     0m0.000s
```

In comparison to the results from the bug report:

| | Before | After |
| :--- | :----: | ---: |
| snow status | 7.0.48s | 2.142s |
| snow add | 48.148s | 31.213s|